Move internal classes to a non exported file
