+++ lib/src/charts/painters/cartesian/cartesian_chart_painter.dart
@@ -2,6 +2,7 @@ import 'dart:math';
 
 import 'package:chart_it/src/charts/constants/defaults.dart';
 import 'package:chart_it/src/charts/data/core.dart';
+import 'package:chart_it/src/charts/data/core/cartesian/cartesian_data_internal.dart';
 import 'package:chart_it/src/charts/painters/cartesian/cartesian_painter.dart';
 import 'package:chart_it/src/charts/painters/text/chart_text_painter.dart';
 import 'package:chart_it/src/charts/state/bar_series_state.dart';
