+++ lib/src/charts/renderers/axis_labels.dart
@@ -1,6 +1,7 @@
 import 'dart:math';
 
 import 'package:chart_it/chart_it.dart';
+import 'package:chart_it/src/charts/data/core/cartesian/cartesian_data_internal.dart';
 import 'package:chart_it/src/charts/renderers/cartesian_scaffold_renderer.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/rendering.dart';
